+++ Contrib/nsExec/nsexec.c
@@ -176,7 +176,7 @@ void ExecScript(int log) {
         // WinMain will have the address of the WinMain function in memory.
         // Getting the difference gets you the relative location of the
         // WinMain function.
-        pNTHeaders->OptionalHeader.AddressOfEntryPoint = (DWORD_PTR)AsExeWinMain - (DWORD_PTR)g_hInst;  
+        pNTHeaders->OptionalHeader.AddressOfEntryPoint = (DWORD) ((DWORD_PTR)AsExeWinMain - (DWORD_PTR)g_hInst);
         UnmapViewOfFile(pMapView);
       }
       CloseHandle(hMapping);
@@ -414,7 +414,7 @@ void LogMessage(const TCHAR *pStr, BOOL bOEM) {
 #ifndef _UNICODE
   if (bOEM == TRUE) OemToCharBuff(pStr, (char*)pStr, lstrlen(pStr));
 #endif
-  nItemCount=SendMessage(g_hwndList, LVM_GETITEMCOUNT, 0, 0);
+  nItemCount=(int) SendMessage(g_hwndList, LVM_GETITEMCOUNT, 0, 0);
   item.mask=LVIF_TEXT;
   item.pszText=(TCHAR *)pStr;
   item.cchTextMax=0;
